+++ example/src/main/scala/example/BasicAuth.scala
@@ -1,7 +1,7 @@
 package example
 
+import zhttp.http.Middleware.basicAuth
 import zhttp.http._
-import zhttp.http.middleware.Middleware.basicAuth
 import zhttp.service.Server
 import zio.{App, ExitCode, URIO}
