+++ example/src/main/scala/example/CSRF.scala
@@ -1,7 +1,7 @@
 package example
 
+import zhttp.http.Middleware.{csrfGenerate, csrfValidate}
 import zhttp.http._
-import zhttp.http.middleware.Middleware.{csrfGenerate, csrfValidate}
 import zhttp.service.Server
 import zio._
